+++ ru.taximaxim.codekeeper.core.tests/src/main/java/ru/taximaxim/codekeeper/core/parsers/ObjReferencesTest.java
@@ -88,7 +88,6 @@ class ObjReferencesTest {
             "numeric",
             "numeric_big",
             "numerology",
-            "object_address",
             "oid",
             "oidjoins",
             "operator",