+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/alter_comments_original.sql
@@ -105,6 +105,10 @@ COMMENT ON VIEW public.test_view IS 'test view';
 
 COMMENT ON COLUMN public.test_view.id IS 'view id col';
 
+CREATE STATISTICS public.s1 ON id, text FROM public.test;
+
+COMMENT ON STATISTICS public.s1 IS 'test statistics';
+
 ALTER TABLE ONLY public.test
     ADD CONSTRAINT test_pkey PRIMARY KEY (id);