+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/chg_owner_diff.sql
@@ -15,3 +15,5 @@ ALTER FUNCTION public.test_fnc(arg character varying) OWNER TO fordfrog;
 ALTER FUNCTION public.trigger_fnc() OWNER TO fordfrog;
 
 ALTER VIEW public.test_view OWNER TO fordfrog;
+
+ALTER STATISTICS public.s1 OWNER TO fordfrog;