+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/depcies/add_statistics_original.sql
@@ -1,0 +1,3 @@
+CREATE SCHEMA first;
+
+CREATE SCHEMA second;