+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/depcies/add_statistics_usr_s2_diff.sql
@@ -1,0 +1,24 @@
+SET search_path = pg_catalog;
+
+-- DEPCY: This FUNCTION is a dependency of STATISTICS: second.s2
+
+CREATE OR REPLACE FUNCTION first.f1(f1 integer) RETURNS integer
+    LANGUAGE plpgsql IMMUTABLE
+    AS $$
+BEGIN
+  SELECT f1 + 1;
+END;
+$$;
+
+CREATE TABLE first.t1 (
+	a integer,
+	b integer
+);
+
+-- DEPCY: This VIEW is a dependency of STATISTICS: second.s2
+
+CREATE MATERIALIZED VIEW first.v1 AS
+	SELECT a, b from first.t1
+WITH DATA;
+
+CREATE STATISTICS second.s2 ON b, first.f1(a) FROM first.v1;