+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/depcies/add_statistics_usr_s2_new.sql
@@ -1,0 +1,5 @@
+CREATE SCHEMA first;
+
+CREATE SCHEMA second;
+
+CREATE STATISTICS second.s2 ON b, first.f1(a) FROM first.v1;