+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/model/difftree/DbObjType.java
@@ -51,6 +51,7 @@ public enum DbObjType {
     INDEX("INDEX"),
     CONSTRAINT("CONSTRAINT"),
     VIEW("VIEW"),
+    STATISTICS("STATISTICS"),
     TRIGGER("TRIGGER"),
     RULE("RULE"),
     POLICY("POLICY");